refactor: FileUtils to common package
