+++ archive-common/src/main/java/site/archive/common/FileUtils.java
@@ -1,4 +1,4 @@
-package site.archive.web.config.security.util;
+package site.archive.common;
 
 import org.springframework.web.multipart.MultipartFile;
